requirements.txt
streamlit>=1.32.0,<2.0.0
openai>=1.30.0,<2.0.0
langchain-core>=0.2.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
